MQ2-20160916

+++ MQ2Labels/ISXEQLabels.cpp
@@ -68,7 +68,7 @@ Id_PMP[] = {
 	// 9999: custom (TooltipReference)
 	{ 9999,   "BrokenParser" },
 
-	{ -1,NULL }
+	{ (DWORD)-1,NULL }
 };
